+++ backend/BucketOfThoughts.Data/Entities/Thoughts/ThoughtDetail.cs
@@ -5,7 +5,7 @@ namespace BucketOfThoughts.Data.Entities;
 public partial class ThoughtDetail : BaseModifiableDbTable
 {
     public string Description { get; set; } = null!;
-    public int ThoughtId { get; set; }
+    public long ThoughtId { get; set; }
     public int? SortOrder { get; set; }
     public virtual Thought Thought { get; set; } = null!;
 }
