+++ backend/BucketOfThoughts.Data/Entities/Thoughts/ThoughtWebsiteLink.cs
@@ -6,8 +6,8 @@ namespace BucketOfThoughts.Data.Entities;
 [Table("ThoughtWebsiteLink")]
 public partial class ThoughtWebsiteLink 
 {
-    public int ThoughtId { get; set; }
-    public int WebsiteLinkId { get; set; }
+    public long ThoughtId { get; set; }
+    public long WebsiteLinkId { get; set; }
     public virtual Thought Thought { get; set; } = null!;
     public virtual WebsiteLink WebsiteLink { get; set; } = null!;
 }
